feat: add unit tests for item and order creation and validation
